requirements.txt
streamlit
ffmpeg-python
Pillow
transformers
numpy<2
torch